name changes, catch exception removed and spport for generic type parameters greater than 9
